Rework some of the autoloader scene-switching and implement tree pausing
Initially the autoloader would delete the world scene before switching to the pause scene,
and similar things occurred when bringing up something like the options menu. This
didn't allow us to pause and resume to what we needed. Therefore, the two
menus, options and pause, were reworked to allow them to overlay on top
of various screens and queue_free themselves, which allowed us to
not delete the world scene, pause its state, and return to gameplay
after pausing right where we left off.